# api-spec.yaml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: >-
      https://virtserver.swaggerhub.com/matthewminish/seng302-2021-api-spec/3.0.1
info:
  description: API specification for the SENG302 2021 project
  version: 3.0.1
  title: Wasteless
  contact:
    email: seng302@canterbury.ac.nz
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
security:
  - CookieAuth:
      - user
      - globalApplicationAdmin
      - defaultGlobalApplicationAdmin
paths:
...
  '/cards':
    post:
      tags:
        - 'UCM2: Card creation'
        - Cards / community marketplace
      summary: Create a new card to go on the community marketplace
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '201':
          description: Card created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  cardId:
                    type: integer
                    description: The ID of the card that has just been created
                    example: 500
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: >-
            Returned if a user who is not a GAA tries to create a card for
            somebody else.
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - type: object
                  required:
                    - creatorId
                  properties:
                    creatorId:
                      type: integer
                      example: 100
                - $ref: '#/components/schemas/Card'
    get:
      tags:
        - 'UCM2: Card creation'
        - Cards / community marketplace
      summary: Retrieve all cards for a given section
      parameters:
        - name: section
          in: query
          description: The section for which cards should be retrieved.
          required: true
          schema:
            type: string
            enum:
              - ForSale
              - Wanted
              - Exchange
        - name: page
          in: query
          description: The page number to retrieve (1 indexed)
          schema:
            type: integer
        - name: resultsPerPage
          in: query
          description: The number of results to show per page
          schema:
            type: integer
        - name: orderBy
          in: query
          description: Card property to order results by
          schema:
            type: string
            enum:
              - lastRenewed
              - created
              - title
              - closes
              - creatorFirstName
              - creatorLastName
              - location
        - name: reverse
          in: query
          description: Whether to flip default sort order (default is ascending)
          schema:
            type: boolean
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: 'Request successful, a (possibly empty) list of cards is returned'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required:
                      - results
                    properties:
                      results:
                        type: array
                        items:
                          $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  '/cards/{id}':
    get:
      tags:
        - 'UCM3: Marketplace section display'
        - Cards / community marketplace
      summary: Retrieve a specific card
      parameters:
        - name: id
          in: path
          description: The id of the card to get.
          required: true
          schema:
            type: integer
            example: 500
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: 'Request successful, a card is returned.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
    put:
      tags:
        - 'UCM9: Edit card'
        - Cards / community marketplace
      summary: Modify an already existing card
      parameters:
        - name: id
          in: path
          description: The id of the card to modify.
          required: true
          schema:
            type: integer
            example: 500
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Card'
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: Card modified successfully
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: >-
            Forbidden: Returned when a user tries to modify a card that they are
            not the creator of AND the user is not a DGAA or GAA
        '406':
          $ref: '#/components/responses/NotAcceptableError'
    delete:
      tags:
        - 'UCM4: Card expiry'
        - Cards / community marketplace
      summary: Delete a community marketplace card
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: The id of the card to get.
          required: true
          schema:
            type: integer
            example: 500
      responses:
        '200':
          description: Card deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: >-
            Forbidden: Returned when a user tries to delete a card that they are
            not the creator of AND the user is not a GAA
        '406':
          $ref: '#/components/responses/NotAcceptableError'
...
  '/cards/{id}/extenddisplayperiod':
    put:
      tags:
        - 'UCM4: Card expiry'
        - Cards / community marketplace
      summary: Extend the display period of a card that is nearing expiry
      parameters:
        - name: id
          in: path
          description: The id of the card to get.
          required: true
          schema:
            type: integer
            example: 500
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: >-
            End of display period successfully extended. Extension amount should
            be a further two weeks.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: >-
            Forbidden: Returned when a user who is neither the creator of the
            card, nor a GAA, tries to extend the display period of the card.
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  '/users/{id}/cards':
    get:
      tags:
        - 'UCM5: Find my cards'
        - Cards / community marketplace
      summary: Retrieve all active cards that a user has created
      parameters:
        - name: id
          in: path
          description: The id of the user for whom to get cards.
          required: true
          schema:
            type: integer
            example: 100
        - name: page
          in: query
          description: The page number to retrieve (1 indexed)
          schema:
            type: integer
        - name: resultsPerPage
          in: query
          description: The number of results to show per page
          schema:
            type: integer
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      responses:
        '200':
          description: 'Request successful, a (possibly empty) list of cards is returned'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required:
                      - results
                    properties:
                      results:
                        type: array
                        items:
                          $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
...
  /cards/search:
    get:
      tags:
        - 'UCM7: Keyword search'
        - Cards / community marketplace
        - Keywords
      summary: Search for cards based on some keywords
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - in: query
          name: keywordIds
          description: The IDs of keywords by which cards are being searched
          schema:
            type: array
            items:
              type: integer
          required: true
        - name: section
          in: query
          description: The section for which cards should be retrieved.
          required: true
          schema:
            type: string
            enum:
              - ForSale
              - Wanted
              - Exchange
        - in: query
          name: union
          description: >-
            Whether or not to get cards that match ANY of the keywords, or only
            cards that match ALL of the keywords. True = Match ANY, False =
            Match ALL
          schema:
            type: boolean
          required: true
        - name: page
          in: query
          description: The page number to retrieve (1 indexed)
          schema:
            type: integer
        - name: resultsPerPage
          in: query
          description: The number of results to show per page
          schema:
            type: integer
        - name: orderBy
          in: query
          description: Card property to order results by
          schema:
            type: string
            enum:
              - lastRenewed
              - created
              - title
              - closes
              - creatorFirstName
              - creatorLastName
              - location
        - name: reverse
          in: query
          description: Whether to flip default sort order (default is ascending)
          schema:
            type: boolean
      responses:
        '200':
          description: 'Request successful, a (possibly empty) list of cards is returned'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required:
                      - results
                    properties:
                      results:
                        type: array
                        items:
                          $ref: '#/components/schemas/Card'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    BadRequestError:
      description: >-
        There was some error with the data supplied by the user, appropriate
        error message(s) should be shown to user
    UnauthorizedError:
      description: Access token is missing or invalid
...
    NotAcceptableError:
      description: >-
        The requested route does exist (so not a 404) but some part of the
        request is not acceptable, for example trying to access a resource by an
        ID that does not exist.
...
  schemas:
    Keyword:
      type: object
      description: Represents a single 'keyword' within the system - similar to a hashtag.
      properties:
        id:
          type: integer
          example: 600
          readOnly: true
          description: 'Unique ID, autogenerated by the system'
        name:
          type: string
          description: The text value of the keyword
          example: Vehicle
        created:
          type: string
          format: date-time
          example: '2021-07-15T05:10:00Z'
          readOnly: true
          description: Date and time when the keyword was created
    Card:
      type: object
      description: Represents an item in the Community Marketplace
      required:
        - title
        - section
        - keywordIds
      properties:
        id:
          type: integer
          example: 500
          readOnly: true
          description: 'Unique ID, autogenerated by the system'
        creator:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/User'
        section:
          type: string
          enum:
            - ForSale
            - Wanted
            - Exchange
        created:
          type: string
          format: date-time
          example: '2021-07-15T05:10:00Z'
          readOnly: true
          description: Date and time when the card was created
        displayPeriodEnd:
          type: string
          format: date-time
          example: '2021-07-29T05:10:00Z'
          readOnly: true
          description: >-
            Date and time when the display period of the card is set to expire.
            This defaults to 2 weeks from creation, and can be extended per
            UCM4.
        title:
          type: string
          description: 'Title of the card, should be suitable for one-line displays.'
          example: 1982 Lada Samara
        description:
          type: string
          description: A more detailed description of the card.
          example: >-
            Beige, suitable for a hen house. Fair condition. Some rust. As is,
            where is. Will swap for budgerigar.
        keywords:
          type: array
          readOnly: true
          description: >-
            Keywords describing the card. A card can have one or more keywords
            associated with it.
          items:
            $ref: '#/components/schemas/Keyword'
        keywordIds:
          type: array
          writeOnly: true
          description: >-
            The IDs of keywords that will be associated with a card once it is
            created. These keywords are created through the POST /keywords
            endpoints (UCM6).
          items:
            type: integer
          example:
            - 20
            - 15
            - 600
...
    Page:
      type: object
      description: Page of results
      required:
        - count
      properties:
        count:
          type: integer
          example: 70
          readOnly: true
          description: Total number of results
...